extract helper functions in tests to utils; move spawn_app() to utils; move test files to tests/api
