+++ tests/api/health_check.rs
@@ -1,0 +1,15 @@
+use crate::utils::spawn_app;
+
+#[tokio::test]
+async fn health_check_works() {
+    let address = spawn_app();
+    let client = reqwest::Client::new();
+
+    let response = client
+        .get(&format!("{address}/health_check"))
+        .send()
+        .await
+        .expect("Failed to execute request");
+    assert!(response.status().is_success());
+    assert_eq!(Some(0), response.content_length());
+}
