+++ tests/api/main.rs
@@ -1,0 +1,2 @@
+pub mod health_check;
+pub mod utils;
