--- tests/health_check_works.rs
@@ -1,26 +1,0 @@
-use actix_web_template::startup::run;
-use std::net::TcpListener;
-
-#[tokio::test]
-async fn health_check_works() {
-    let address = spawn_app();
-    let client = reqwest::Client::new();
-
-    let response = client
-        .get(&format!("{address}/health_check"))
-        .send()
-        .await
-        .expect("Failed to execute request");
-    assert!(response.status().is_success());
-    assert_eq!(Some(0), response.content_length());
-}
-
-// Spawn an instance of the app using a random available port and return the
-// address used, including the selected port.
-fn spawn_app() -> String {
-    let listener = TcpListener::bind("127.0.0.1:0").expect("Failed to bind to random port");
-    let port = listener.local_addr().unwrap().port();
-    let server = run(listener).expect("Failed to bind address");
-    let _ = tokio::spawn(server);
-    format!("http://127.0.0.1:{port}")
-}
